AP_RangeFinder: Small optimizations

+++ libraries/AP_RangeFinder/AP_RangeFinder_Backend_CAN.h
@@ -32,7 +32,7 @@ protected:
     virtual bool handle_frame(AP_HAL::CANFrame &frame) = 0;
 
     // maximum time between readings before we change state to NoData:
-    virtual uint16_t read_timeout_ms() const { return 200; }
+    virtual uint32_t read_timeout_ms() const { return 200; }
 
     virtual MAV_DISTANCE_SENSOR _get_mav_distance_sensor_type() const override {
         return MAV_DISTANCE_SENSOR_RADAR;
